\name{bivpois-package}
\alias{bivpois-package}
\docType{package}
\title{
Bivariate Poisson Distribution
}
\description{
The Bivariate Poisson Distribution.
}
\details{
\tabular{lll}{
Package: \tab bivpois\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-10-18\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Kawamura K. (1984). Direct calculation of maximum likelihood estimator for the bivariate
Poisson distribution. Kodai Mathematical Journal, 7(2): 211--221.

Kocherlakota S. and Kocherlakota K. (1998). Bivariate discrete distributions. Wiley Online
Library.

Karlis D. and Ntzoufras I. (2003). Analysis of sports data by using bivariate poisson models.
Journal of the Royal Statistical Society: Series D (The Statistician), 52(3): 381--393.
}



