\name{splitbeta}
\alias{splitbeta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal Function of BIVPOIS Package}
\description{ 
This function was made only for internal use in Bivpois package and it should not be called separately
}
\author{
1.      Dimitris Karlis, Department of Statistics, Athens University of Economics and Business, Athens, Greece, \email{karlis@aueb.gr} .

2.      Ioannis Ntzoufras, Department of Statistics, Athens University of Economics and Business, Athens, Greece, \email{ntzoufras@aueb.gr}  .
}
\keyword{internal}
